+++ g2g-plugins/src/identity.rs
@@ -65,6 +65,9 @@ impl AsyncElement for IdentityTransform {
                 PipelinePacket::CapsChanged(c) => {
                     out.push(PipelinePacket::CapsChanged(c)).await?;
                 }
+                PipelinePacket::Flush => {
+                    out.push(PipelinePacket::Flush).await?;
+                }
                 PipelinePacket::Eos => {}
             }
             Ok(())